Struct google_dfareporting2d6::CreativeAssetMetadata

pub struct CreativeAssetMetadata {
    pub id_dimension_value: Option<DimensionValue>,
    pub kind: Option<String>,
    pub detected_features: Option<Vec<String>>,
    pub click_tags: Option<Vec<ClickTag>>,
    pub warned_validation_rules: Option<Vec<String>>,
    pub asset_identifier: Option<CreativeAssetId>,
    pub id: Option<String>,
}

CreativeAssets contains properties of a creative asset file which will be uploaded or has already been uploaded. Refer to the creative sample code for how to upload assets and insert a creative.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert creative assets (request|response)

Fields
id_dimension_value: Option<DimensionValue>

Dimension value for the numeric ID of the asset. This is a read-only, auto-generated field.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#creativeAssetMetadata".

detected_features: Option<Vec<String>>

List of feature dependencies for the creative asset that are detected by DCM. Feature dependencies are features that a browser must be able to support in order to render your HTML5 creative correctly. This is a read-only, auto-generated field.

click_tags: Option<Vec<ClickTag>>

List of detected click tags for assets. This is a read-only auto-generated field.

warned_validation_rules: Option<Vec<String>>

Rules validated during code generation that generated a warning. This is a read-only, auto-generated field.

Possible values are: - "ADMOB_REFERENCED" - "ASSET_FORMAT_UNSUPPORTED_DCM" - "ASSET_INVALID" - "CLICK_TAG_HARD_CODED" - "CLICK_TAG_INVALID" - "CLICK_TAG_IN_GWD" - "CLICK_TAG_MISSING" - "CLICK_TAG_MORE_THAN_ONE" - "CLICK_TAG_NON_TOP_LEVEL" - "COMPONENT_UNSUPPORTED_DCM" - "ENABLER_UNSUPPORTED_METHOD_DCM" - "EXTERNAL_FILE_REFERENCED" - "FILE_DETAIL_EMPTY" - "FILE_TYPE_INVALID" - "GWD_PROPERTIES_INVALID" - "HTML5_FEATURE_UNSUPPORTED" - "LINKED_FILE_NOT_FOUND" - "MAX_FLASH_VERSION_11" - "MRAID_REFERENCED" - "NOT_SSL_COMPLIANT" - "ORPHANED_ASSET" - "PRIMARY_HTML_MISSING" - "SVG_INVALID" - "ZIP_INVALID"

asset_identifier: Option<CreativeAssetId>

ID of the creative asset. This is a required field.

id: Option<String>

Numeric ID of the asset. This is a read-only, auto-generated field.