Struct google_dfareporting2d6::SiteListCall

pub struct SiteListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Retrieves a list of sites, possibly filtered.

A builder for the list method supported by a site resource. It is not used directly, but through a SiteMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.sites().list("profileId")
             .unmapped_site(true)
             .subaccount_id("et")
             .sort_order("aliquyam")
             .sort_field("nonumy")
             .search_string("sit")
             .page_token("aliquyam")
             .max_results(-95)
             .add_ids("magna")
             .add_directory_site_ids("gubergren")
             .add_campaign_ids("sit")
             .approved(false)
             .ad_words_site(true)
             .accepts_publisher_paid_placements(false)
             .accepts_interstitial_placements(true)
             .accepts_in_stream_video_placements(false)
             .doit();

Methods

impl<'a, C, A> SiteListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, SitesListResponse)>

Perform the operation you have build so far.

fn profile_id(self, new_value: &str) -> SiteListCall<'a, C, A>

User profile ID associated with this request.

Sets the profile id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn unmapped_site(self, new_value: bool) -> SiteListCall<'a, C, A>

Select only sites that have not been mapped to a directory site.

Sets the unmapped site query property to the given value.

fn subaccount_id(self, new_value: &str) -> SiteListCall<'a, C, A>

Select only sites with this subaccount ID.

Sets the subaccount id query property to the given value.

fn sort_order(self, new_value: &str) -> SiteListCall<'a, C, A>

Order of sorted results, default is ASCENDING.

Sets the sort order query property to the given value.

fn sort_field(self, new_value: &str) -> SiteListCall<'a, C, A>

Field by which to sort the list.

Sets the sort field query property to the given value.

fn search_string(self, new_value: &str) -> SiteListCall<'a, C, A>

Allows searching for objects by name, ID or keyName. Wildcards (*) are allowed. For example, "site*2015" will return objects with names like "site June 2015", "site April 2015", or simply "site 2015". Most of the searches also add wildcards implicitly at the start and the end of the search string. For example, a search string of "site" will match objects with name "my site", "site 2015", or simply "site".

Sets the search string query property to the given value.

fn page_token(self, new_value: &str) -> SiteListCall<'a, C, A>

Value of the nextPageToken from the previous result page.

Sets the page token query property to the given value.

fn max_results(self, new_value: i32) -> SiteListCall<'a, C, A>

Maximum number of results to return.

Sets the max results query property to the given value.

fn add_ids(self, new_value: &str) -> SiteListCall<'a, C, A>

Select only sites with these IDs.

Append the given value to the ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn add_directory_site_ids(self, new_value: &str) -> SiteListCall<'a, C, A>

Select only sites with these directory site IDs.

Append the given value to the directory site ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn add_campaign_ids(self, new_value: &str) -> SiteListCall<'a, C, A>

Select only sites with these campaign IDs.

Append the given value to the campaign ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn approved(self, new_value: bool) -> SiteListCall<'a, C, A>

Select only approved sites.

Sets the approved query property to the given value.

fn ad_words_site(self, new_value: bool) -> SiteListCall<'a, C, A>

Select only AdWords sites.

Sets the ad words site query property to the given value.

fn accepts_publisher_paid_placements(self, new_value: bool) -> SiteListCall<'a, C, A>

Select only sites that accept publisher paid placements.

Sets the accepts publisher paid placements query property to the given value.

fn accepts_interstitial_placements(self, new_value: bool) -> SiteListCall<'a, C, A>

This search filter is no longer supported and will have no effect on the results returned.

Sets the accepts interstitial placements query property to the given value.

fn accepts_in_stream_video_placements(self, new_value: bool) -> SiteListCall<'a, C, A>

This search filter is no longer supported and will have no effect on the results returned.

Sets the accepts in stream video placements query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> SiteListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> SiteListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> SiteListCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Dfatrafficking.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.