Struct google_dfareporting2d6::SiteMethods

pub struct SiteMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on site resources. It is not used directly, but through the Dfareporting hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_dfareporting2d6 as dfareporting2d6;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dfareporting2d6::Dfareporting;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dfareporting::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.sites();

Methods

impl<'a, C, A> SiteMethods<'a, C, A>

fn insert(&self, request: Site, profile_id: &str) -> SiteInsertCall<'a, C, A>

Create a builder to help you perform the following task:

Inserts a new site.

Arguments

request - No description provided.

profileId - User profile ID associated with this request.

fn get(&self, profile_id: &str, id: &str) -> SiteGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets one site by ID.

Arguments

profileId - User profile ID associated with this request.

id - Site ID.

fn list(&self, profile_id: &str) -> SiteListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of sites, possibly filtered.

Arguments

profileId - User profile ID associated with this request.

fn update(&self, request: Site, profile_id: &str) -> SiteUpdateCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing site.

Arguments

request - No description provided.

profileId - User profile ID associated with this request.

fn patch(&self, request: Site, profile_id: &str, id: &str) -> SitePatchCall<'a, C, A>

Create a builder to help you perform the following task:

Updates an existing site. This method supports patch semantics.

Arguments

request - No description provided.

profileId - User profile ID associated with this request.

id - Site ID.